+++ lib/model/app/contributors.dart
@@ -83,6 +83,10 @@ class Contributors {
     Contributor('wanzh', 'wanzh66666@gmail.com'),
     Contributor('ID J', 'tabby4442@gmail.com'),
     Contributor('randint', 'lancameb@hotmail.com'),
+    Contributor('امیر جهانگرد', 'ijahangard.a@gmail.com'),
+    Contributor('slasb37', 'p84haghi@gmail.com'),
+    Contributor('mimvahedi', 'vahedi0vahedi@gmail.com'),
+    Contributor('Alireza Rashidi', 'alirezarashidigoorabi@gmail.com'),
     // Contributor('Alvi Khan', 'aveenalvi@gmail.com'), // Bengali
     // Contributor('Htet Oo Hlaing', 'htetoh2006@outlook.com'), // Burmese
     // Contributor('Khant', 'khant@users.noreply.hosted.weblate.org'), // Burmese
@@ -96,10 +100,6 @@ class Contributors {
     // Contributor('Rasti K5', 'rasti.khdhr@gmail.com'), // Kurdish (Central)
     // Contributor('Raman', 'xysed@tutanota.com'), // Malayalam
     // Contributor('Subham Jena', 'subhamjena8465@gmail.com'), // Odia
-    // Contributor('امیر جهانگرد', 'ijahangard.a@gmail.com'), // Persian
-    // Contributor('slasb37', 'p84haghi@gmail.com'), // Persian
-    // Contributor('mimvahedi', 'vahedi0vahedi@gmail.com'), // Persian
-    // Contributor('Alireza Rashidi', 'alirezarashidigoorabi@gmail.com'), // Persian
     // Contributor('Prasanta-Hembram', 'Prasantahembram720@gmail.com'), // Santali
     // Contributor('mytja', 'mamnju21@gmail.com'), // Slovenian
     // Contributor('Shift18', 'bribable.lawyer@posteo.net'), // Swedish